synch before moving to git
git-svn-id: https://sub.unidata.ucar.edu/common/utils/trunk@14739 2e3060e4-1ed8-cf1f-9cc2-9586e844a556

+++ templates/bytes_template.h
@@ -6,8 +6,8 @@
 
 typedef struct XXbytes {
   int nonextendible; /* 1 => fail if an attempt is made to extend this buffer*/
-  unsigned int alloc;
-  unsigned int length;
+  unsigned long alloc;
+  unsigned long length;
   char* content;
 } XXbytes;
 
@@ -19,8 +19,8 @@ typedef struct XXbytes {
 
 EXTERNC XXbytes* xxbytesnew(void);
 EXTERNC void xxbytesfree(XXbytes*);
-EXTERNC int xxbytessetalloc(XXbytes*,unsigned int);
-EXTERNC int xxbytessetlength(XXbytes*,unsigned int);
+EXTERNC int xxbytessetalloc(XXbytes*,unsigned long);
+EXTERNC int xxbytessetlength(XXbytes*,unsigned long);
 EXTERNC int xxbytesfill(XXbytes*, char fill);
 
 /* Produce a duplicate of the contents*/
@@ -29,14 +29,14 @@ EXTERNC char* xxbytesdup(XXbytes*);
 EXTERNC char* xxbytesextract(XXbytes*);
 
 /* Return the ith byte; -1 if no such index */
-EXTERNC int xxbytesget(XXbytes*,unsigned int);
+EXTERNC int xxbytesget(XXbytes*,unsigned long);
 /* Set the ith byte */
-EXTERNC int xxbytesset(XXbytes*,unsigned int,char);
+EXTERNC int xxbytesset(XXbytes*,unsigned long,char);
 
 /* Append one byte */
 EXTERNC int xxbytesappend(XXbytes*,char); /* Add at Tail */
 /* Append n bytes */
-EXTERNC int xxbytesappendn(XXbytes*,const void*,unsigned int); /* Add at Tail */
+EXTERNC int xxbytesappendn(XXbytes*,const void*,unsigned long); /* Add at Tail */
 
 /* Null terminate the byte string without extending its length (for debugging) */
 EXTERNC int xxbytesnull(XXbytes*);
@@ -45,14 +45,14 @@ EXTERNC int xxbytesnull(XXbytes*);
 EXTERNC int xxbytescat(XXbytes*,const char*);
 
 /* Set the contents of the buffer; mark the buffer as non-extendible */
-EXTERNC int xxbytessetcontents(XXbytes*, char*, unsigned int);
+EXTERNC int xxbytessetcontents(XXbytes*, char*, unsigned long);
 
 /* Following are always "in-lined"*/
-#define xxbyteslength(bb) ((bb)!=NULL?(bb)->length:0U)
-#define xxbytesalloc(bb) ((bb)!=NULL?(bb)->alloc:0U)
+#define xxbyteslength(bb) ((bb)!=NULL?(bb)->length:0)
+#define xxbytesalloc(bb) ((bb)!=NULL?(bb)->alloc:0)
 #define xxbytescontents(bb) (((bb)!=NULL && (bb)->content!=NULL)?(bb)->content:(char*)"")
 #define xxbytesextend(bb,len) xxbytessetalloc((bb),(len)+(bb->alloc))
-#define xxbytesclear(bb) ((bb)!=NULL?(bb)->length=0:0U)
-#define xxbytesavail(bb,n) ((bb)!=NULL?((bb)->alloc - (bb)->length) >= (n):0U)
+#define xxbytesclear(bb) ((bb)!=NULL?(bb)->length=0:0)
+#define xxbytesavail(bb,n) ((bb)!=NULL?((bb)->alloc - (bb)->length) >= (n):0)
 
 #endif /*XXBYTES_H*/
